+++ src/lib/pages/proposal-details/components/index.ts
@@ -1,0 +1,1 @@
+export * from "./ProposalTop";
